--- include/pose/DlibPoser.h
@@ -1,36 +1,0 @@
-#ifndef PROJECT_WG_DLIB_POSER
-#define PROJECT_WG_DLIB_POSER
-
-#include <utility>
-
-#include "Pose.h"
-#include "util/Derive.h"
-
-namespace cv {
-class Mat;
-} // namespace cv
-
-namespace cw {
-
-class DlibPoserPrivate;
-
-class DlibPoser {
-public:
-  DlibPoser();
-  ~DlibPoser();
-
-  bool Open();
-
-  std::pair<bool, HeadPose>
-  EstimateHeadPose(cv::Mat const& frame, bool useEyeLandmarks);
-
-  CW_DERIVE_UNCOPYABLE(DlibPoser)
-  CW_DERIVE_UNMOVABLE(DlibPoser)
-
-private:
-  DlibPoserPrivate *m_Private;
-};
-
-} // namespace cw
-
-#endif // PROJECT_WG_DLIB_POSER
